chore: flow report tests
